Remove dep cycle between constants and types
