+++ source/transfers/package.json
@@ -9,15 +9,15 @@
   },
   "scripts": {
     "clean": "rm -rf ./build",
-    "compile": "truffle compile",
+    "compile": "yarn clean; truffle compile",
     "coverage": "node_modules/.bin/solidity-coverage",
     "ganache": "ganache-cli -p 8545 --gasLimit 0xfffffffffff --time '2017-05-10T00:00:00+00:00'",
     "hint": "yarn solhint \"./contracts/**/*.sol\"",
     "lint": "yarn eslint \"./test/**/*.js\"",
-    "test": "truffle test"
+    "test": "yarn compile; truffle test"
   },
   "devDependencies": {
-    "@airswap/constants": "0.3.5",
+    "@airswap/constants": "0.3.6",
     "@airswap/test-utils": "0.1.6",
     "@airswap/tokens": "0.1.4",
     "solidity-coverage": "^0.6.3"
