+++ source/wrapper/package.json
@@ -15,14 +15,14 @@
   },
   "scripts": {
     "clean": "rm -rf ./build",
-    "compile": "truffle compile",
+    "compile": "yarn clean; truffle compile",
     "coverage": "node_modules/.bin/solidity-coverage",
     "ganache": "ganache-cli -p 8545 --gasLimit 0xfffffffffff --time '2017-05-10T00:00:00+00:00' -s 0",
     "hint": "yarn solhint \"./contracts/**/*.sol\"",
     "lint": "yarn eslint \"./test/**/*.js\"",
     "truffle": "node_modules/.bin/truffle --network development",
     "cp_migration_flat": "mkdir -p flatten/; cp contracts/Migrations.sol flatten/",
-    "test": "truffle test",
+    "test": "yarn compile; truffle test",
     "flatten": "truffle run flatten",
     "migrate": "yarn cp_migration_flat; truffle migrate --skip-dry-run",
     "verify": "truffle run verify"
@@ -33,9 +33,10 @@
     "@airswap/tokens": "0.1.4"
   },
   "devDependencies": {
-    "@airswap/constants": "0.3.5",
+    "@airswap/constants": "0.3.6",
     "@airswap/indexer": "3.6.9",
     "@airswap/test-utils": "0.1.6",
+    "@airswap/types": "3.5.11",
     "@airswap/utils": "0.3.11",
     "ethers": "^4.0.45",
     "solidity-coverage": "^0.6.3"
