+++ tools/metadata/package.json
@@ -25,8 +25,8 @@
     "test": "yarn compile; mocha -r ts-node/register test/**/*.ts"
   },
   "dependencies": {
-    "@airswap/constants": "0.3.5",
-    "@airswap/types": "3.5.10",
+    "@airswap/constants": "0.3.6",
+    "@airswap/types": "3.5.11",
     "axios": "^0.19.2",
     "eth-contract-metadata": "^1.12.1",
     "ethers": "^4.0.45"
