+++ tools/protocols/package.json
@@ -24,12 +24,12 @@
     "test": "yarn compile; mocha -r ts-node/register test/**/*.ts"
   },
   "dependencies": {
-    "@airswap/constants": "0.3.5",
+    "@airswap/constants": "0.3.6",
     "@airswap/delegate": "2.6.9",
     "@airswap/indexer": "3.6.9",
     "@airswap/swap": "5.4.8",
     "@airswap/tokens": "0.1.4",
-    "@airswap/types": "3.5.10",
+    "@airswap/types": "3.5.11",
     "@airswap/utils": "0.3.11",
     "@airswap/validator": "1.2.6",
     "bignumber.js": "^9.0.0",
