+++ tools/utils/package.json
@@ -24,8 +24,8 @@
     "test": "yarn compile; mocha -r ts-node/register test/**/*.ts"
   },
   "dependencies": {
-    "@airswap/constants": "0.3.5",
-    "@airswap/types": "3.5.10",
+    "@airswap/constants": "0.3.6",
+    "@airswap/types": "3.5.11",
     "@types/ethereumjs-abi": "^0.6.3",
     "@types/ethereumjs-util": "^6.1.0",
     "ethereumjs-abi": "^0.6.8",
